# Improve decoding throughput