--- a/tests/molecule_benchmark_test.go
+++ b/tests/molecule_benchmark_test.go
@@ -2,19 +2,18 @@ package moleculetest
 
 import (
 	"testing"
-	"time"
 
 	"github.com/richardartoul/molecule"
 	"github.com/richardartoul/molecule/src/codec"
-	"github.com/richardartoul/molecule/src/proto"
+	simple "github.com/richardartoul/molecule/src/proto"
 
 	"github.com/golang/protobuf/proto"
-	"github.com/google/gofuzz"
+	fuzz "github.com/google/gofuzz"
 )
 
 func BenchmarkMolecule(b *testing.B) {
 	var (
-		seed   = time.Now().UnixNano()
+		seed   = int64(1623963202)
 		fuzzer = fuzz.NewWithSeed(seed)
 	)
 	// Limit slice size to prevent tests from taking too long.
@@ -31,6 +30,103 @@ func BenchmarkMolecule(b *testing.B) {
 		for i := 0; i < b.N; i++ {
 			err := proto.Unmarshal(marshaled, into)
 			noErr(err)
+		}
+	})
+
+	b.Run("unmarshal all", func(b *testing.B) {
+		msgBuffer := codec.NewBuffer(marshaled)
+		m := &simple.Simple{}
+		for i := 0; i < b.N; i++ {
+			msgBuffer.Reset(marshaled)
+			err := molecule.MessageEach(msgBuffer, func(fieldNum int32, value molecule.Value) (bool, error) {
+				var err error
+				switch fieldNum {
+				case 1:
+					m.Double, err = value.AsDouble()
+				case 2:
+					m.Float, err = value.AsFloat()
+				case 3:
+					m.Int32, err = value.AsInt32()
+				case 4:
+					m.Int64, err = value.AsInt64()
+				case 5:
+					m.Uint32, err = value.AsUint32()
+				case 6:
+					m.Uint64, err = value.AsUint64()
+				case 7:
+					m.Sint32, err = value.AsSint32()
+				case 8:
+					m.Sint64, err = value.AsSint64()
+				case 9:
+					m.Fixed32, err = value.AsFixed32()
+				case 10:
+					m.Fixed64, err = value.AsFixed64()
+				case 11:
+					m.Sfixed32, err = value.AsSFixed32()
+				case 12:
+					m.Sfixed64, err = value.AsSFixed64()
+				case 13:
+					m.Bool, err = value.AsBool()
+				case 14:
+					m.String_, err = value.AsStringUnsafe()
+				case 15:
+					m.Bytes, err = value.AsBytesUnsafe()
+				case 16:
+
+				}
+
+				return err == nil, err
+			})
+			noErr(err)
+		}
+	})
+
+	b.Run("unmarshal loop", func(b *testing.B) {
+		msgBuffer := codec.NewBuffer(marshaled)
+		m := &simple.Simple{}
+		for i := 0; i < b.N; i++ {
+			msgBuffer.Reset(marshaled)
+			value := molecule.Value{}
+			for !msgBuffer.EOF() {
+				fieldNum, err := molecule.Next(msgBuffer, &value)
+				noErr(err)
+
+				switch fieldNum {
+				case 1:
+					m.Double, err = value.AsDouble()
+				case 2:
+					m.Float, err = value.AsFloat()
+				case 3:
+					m.Int32, err = value.AsInt32()
+				case 4:
+					m.Int64, err = value.AsInt64()
+				case 5:
+					m.Uint32, err = value.AsUint32()
+				case 6:
+					m.Uint64, err = value.AsUint64()
+				case 7:
+					m.Sint32, err = value.AsSint32()
+				case 8:
+					m.Sint64, err = value.AsSint64()
+				case 9:
+					m.Fixed32, err = value.AsFixed32()
+				case 10:
+					m.Fixed64, err = value.AsFixed64()
+				case 11:
+					m.Sfixed32, err = value.AsSFixed32()
+				case 12:
+					m.Sfixed64, err = value.AsSFixed64()
+				case 13:
+					m.Bool, err = value.AsBool()
+				case 14:
+					m.String_, err = value.AsStringUnsafe()
+				case 15:
+					m.Bytes, err = value.AsBytesUnsafe()
+				case 16:
+
+				}
+				noErr(err)
+			}
 		}
 	})