This commit combines a few changes that improve the decoding throughput.

```
name                                         old time/op  new time/op  delta
Molecule/standard_unmarshal-8                1.28µs ± 8%  1.32µs ± 4%   +2.64%  (p=0.000 n=23+23)
Molecule/unmarshal_all-8                      330ns ± 2%   211ns ± 6%  -36.09%  (p=0.000 n=25+21)
Molecule/unmarshal_loop-8                     322ns ± 1%   169ns ± 3%  -47.62%  (p=0.000 n=22+22)
Molecule/unmarshal_single_with_molecule-8     257ns ± 2%   165ns ± 2%  -35.72%  (p=0.000 n=21+21)
Molecule/unmarshal_multiple_with_molecule-8  1.49µs ± 7%  0.99µs ± 4%  -33.36%  (p=0.000 n=25+24)
```

I attempted to switch one of our apps to use molecule instead of the gogo generated code but I found that the CPU usage of using molecule significantly worse.

The most obvious thing in the profiler was `DecodeVarint` so I initially focused my efforts there.  I originally replaced `DecodeVarint` with the body of `decodeVarintSlow` as this allowed the function to be inlined.  However, the implementation from https://github.com/dennwc/varint was faster still, even accounting for the fact that it is too complex to be inlined.

The improvements here mainly focus on either inlining methods directly or modifying them so that they are eligible to be inlined.  My intuition is that the call-site overhead was actually contributing to the the CPU usage so inlining where possible seemed to help.

I also added a `Next()` method that allows a different method of iterating over the values, you can see in the benchmark that this is significantly faster than using `MessageEach`, again I think this is due to call-site overhead / stack management.  I'm not sure how you feel about having competing APIs like this but the improvement to me seems worthwhile.

Open to suggestions on how to improve this further.